+++ src/persistent_store.rs
@@ -1,4 +1,4 @@
-use crate::storage::Storage;
+
 use crate::types::BackgroundService;
 use async_trait::async_trait;
 use fjall::{Keyspace, KeyspaceCreateOptions, PersistMode, Slice};