+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Changelog
 * Preparation to full implementation of cache invalidation. Refactored key serialization for persistent storage. *
   *BREAKING:** Need to
   drop existing persistent storage before update (actually it's not breaking, but lead to dead weight in persistent
-  cache).
+  cache)
+* Add env var to restrict max output size(height, width)
 
 0.1.1
 --------